+++ src/opts.rs
@@ -52,6 +52,14 @@ pub struct Options {
     /// Whether or not to show articles marked as read by Newsboat
     #[serde(default = "default_bool::<true>")]
     pub show_read_articles: bool,
+    /// Define whether to scrape reddit rss links
+    #[serde(default = "default_bool::<true>")]
+    pub scrape_reddit_links: bool,
+    /// If set to true Liveboat will scrape all external
+    /// Hackers News links and subsitute the content with
+    /// retrieved data.
+    #[serde(default = "default_bool::<true>")]
+    pub scrape_hn_links: bool,
     /// Define whether or not to include article content in generated
     /// rss feeds (might increase size significantly)
     #[serde(default = "default_bool::<true>")]
@@ -89,8 +97,10 @@ impl Options {
             site_path: default_site_path(),
             show_read_articles: true,
             template_name: default_template_name(),
-            include_article_content_in_rss_feeds: false,
+            include_article_content_in_rss_feeds: true,
             time_threshold: 20,
+            scrape_reddit_links: true,
+            scrape_hn_links: true,
             newsboat_urls_file: default_newsboat_url_file(),
             newsboat_cache_file: default_newsboat_cache_file(),
             build_dir: default_build_dir(),