+++ templates/default/src/include/src/components/FeedList.vue
@@ -1,11 +1,13 @@
 <script setup>
-import FeedItem from './FeedItem.vue'
+import { ref } from 'vue'
+import FeedItems from './FeedItems.vue'
 import AudioPlayer from './AudioPlayer.vue'
 import FilterBox from './FilterBox.vue'
 import EmbedModal from './EmbedModal.vue'
 import { useFeedsStore } from '@/stores/feeds'
 import { useEmbedStore } from '@/stores/embed'
 import { useAudioStore } from '@/stores/audio'
+import { useFiltersStore } from '@/stores/filters'
 import { storeToRefs } from 'pinia'
 
 const props = defineProps({
@@ -19,17 +21,83 @@ const props = defineProps({
   },
 })
 
+const expandedFeed = ref(null)
+const expandedArticles = ref([])
 const embedStore = useEmbedStore()
 const audioStore = useAudioStore()
 const feedsStore = useFeedsStore()
+const filterStore = useFiltersStore()
+
 const { feeds } = storeToRefs(feedsStore)
+
+const handleFeedExpand = async (expandData) => {
+  expandData.then((res) => {
+    expandedFeed.value = res.feedId
+    expandedArticles.value = res.articleIds
+  })
+}
+
+const handleFeedUnexpand = () => {
+  expandedFeed.value = null
+  expandedArticles.value = []
+}
+
+const handleArticleExpand = (articleId) => {
+  expandedArticles.value.push(articleId)
+}
+
+const handleArticleUnexpand = (articleId) => {
+  expandedArticles.value = expandedArticles.value.filter((i) => {
+    return i !== articleId
+  })
+}
+
+const showExpandedFeed = (feed) => {
+  if (expandedFeed.value == null) {
+    return false
+  }
+  return feed.id === expandedFeed.value
+}
+
+const generateFirehoseFeed = () => {
+  return {
+    id: null,
+    url: null,
+    title: 'Firehose',
+    displayTitle: 'Firehose',
+    itemCount: 0,
+    isQuery: false,
+  }
+}
 </script>
 
 <template>
   <FilterBox />
-  <div class="feed-list-wrapper" v-for="feed in feeds" :key="feed.id">
+  <div v-if="filterStore.firehose">
+    <FeedItems
+      :feed="generateFirehoseFeed()"
+      :filtered="props.filtered"
+      :archived="props.archived"
+      :firehose="true"
+      :expandedArticles="expandedArticles"
+      @expand-article="handleArticleExpand"
+      @unexpand-article="handleArticleUnexpand"
+    />
+  </div>
+  <div class="feed-list-wrapper" v-else v-for="feed in feeds" :key="feed.id">
     <Transition>
-      <FeedItem :feed="feed" :filtered="props.filtered" :archived="props.archived"> </FeedItem>
+      <FeedItems
+        :feed="feed"
+        :filtered="props.filtered"
+        :archived="props.archived"
+        :expand="showExpandedFeed(feed)"
+        :firehose="false"
+        :expandedArticles="expandedArticles"
+        @expand-feed="handleFeedExpand"
+        @unexpand-feed="handleFeedUnexpand"
+        @expand-article="handleArticleExpand"
+        @unexpand-article="handleArticleUnexpand"
+      />
     </Transition>
   </div>
   <EmbedModal