+++ templates/default/src/include/src/components/FilterBox.vue
@@ -8,13 +8,22 @@ const setTimeLimit = (limit) => {
   fStore.$patch({
     daysBackCount: limit,
     filterByDays: true,
+    firehose: false,
   })
 }
 const setItemLimit = (limit) => {
   fStore.$patch({
     itemCount: limit,
     filterByDays: false,
+    firehose: false,
   })
+}
+const setFirehose = () => {
+  fStore.$patch({
+    firehose: true,
+    filterByDays: false,
+  })
+
 }
 
 const filters = ref(fStore.filters)
@@ -43,15 +52,15 @@ watch(searchFeedsTerm, (val) => {
   <div class="filter-container">
     <span class="filter-box"
       ><button
-        :class="{ selected: filters.daysBackCount === 7 && filters.filterByDays === true }"
-        @click="setTimeLimit(7)"
+        :class="{ selected: filters.firehose === true }"
+        @click="setFirehose()"
       >
-        Last week
+      Firehose
       </button></span
     >
     <span class="filter-box"
       ><button
-        :class="{ selected: filters.daysBackCount === 1 && filters.filterByDays === true }"
+        :class="{ selected: filters.daysBackCount === 1 && filters.filterByDays === true && !filters.firehose }"
         @click="setTimeLimit(1)"
       >
         Last day
@@ -59,7 +68,7 @@ watch(searchFeedsTerm, (val) => {
     >
     <span class="filter-box"
       ><button
-        :class="{ selected: filters.itemCount === 50 && filters.filterByDays === false }"
+        :class="{ selected: filters.itemCount === 50 && filters.filterByDays === false && !filters.firehose }"
         @click="setItemLimit(50)"
       >
         Last 50
@@ -67,7 +76,7 @@ watch(searchFeedsTerm, (val) => {
     >
     <span class="filter-box"
       ><button
-        :class="{ selected: filters.itemCount === 20 && filters.filterByDays === false }"
+        :class="{ selected: filters.itemCount === 20 && filters.filterByDays === false && !filters.firehose }"
         @click="setItemLimit(20)"
       >
         Last 20